+++ backend/collaboration-service/src/controllers/handlers/socket-handler.ts
