+++ backend/collaboration-service/src/controllers/index.ts
@@ -1,1 +1,1 @@
-export * from "./handlers/socker-handler";
+export * from "./handlers/socket-handler";
